Update model to include jur.Pers.Mitgl. and update KalteBetr.Rechn.
JuristischePerson are now able to include Kontakte as Mitglieder.
For this a JoinTable is created, because it is a many to many relation.

KalteBetriebskostenRechnung is no longer refering to the KalteBetriebskostenpunkt, because they may change.
The Rechnung references the Adresse and Typ, which should be closer to reality and is more fitting if things change.