Introduce warme Kosten to Betriebskostenabrechnung
For now only direkte Kosten are considered.
The first steps were to make a distinction between kalte and warme Betriebskosten.

The next step is to introduce warme Betriebskosten to the extend of the whole group.
Some complications have to be overcome, since a lot of variables can be set for this very calculation.