feat: Verwalter connects Account and Wohnung
Verwalter is not connected to Kontakt anymore.
Therefore every Account is connected directly to the Wohungen.
Accounts still have join Tables with Kontakte, but that is used
to create suggestions to the owner or the admin to set Verwalter in the
frontend for the owner or the admin.

This setup results in the Account having a multitude of Verwalter
connections to individual Wohnungen, but this
is a frontend problem and can be solved by grouping using
the Kontakte Join Table.

This allows for a more transparent authorization and more granular
control for the respective owner/admin.