feat: Add Owner UserRole and VerwalterRolle
This Role is used to determine whether a User is authorized to create
new Wohnungen.
The User that is creating a new Wohnung will get a Verwalter link
to the Wohnung with VerwalterRolle Eigentuemer.

Eigentuemer is added as VerwalterRolle.
The Eigentuemer is used to create new Verwalter with VerwalterRolle
Vollmacht.